xong hết

+++ QuanLyNhaTro/DanhSachONhaTro.cs
@@ -38,36 +38,65 @@ namespace QuanLyNhaTro
             dGVDSOTro.DataSource = dao.loadthuePhong();
           
         }
-
+        public void anhxa()
+        {
+            txtTienDatCoc.Text = "";
+            txtHoTen.Text = "";
+            txtCMND.Text = "";
+            txtSoDienThoai.Text = "";
+            txtDiaChi.Text = "";
+            txtTienDatCoc.Focus();
+        }
         private void dGVDSOTro_CellClick(object sender, DataGridViewCellEventArgs e)
         {
-            int i = dGVDSOTro.CurrentRow.Index;
-            txtTienDatCoc.Text = dGVDSOTro.Rows[i].Cells[1].Value.ToString();
-            dTPNgayThue.Text = dGVDSOTro.Rows[i].Cells[2].Value.ToString();
-            dTPNgayTra.Text = dGVDSOTro.Rows[i].Cells[3].Value.ToString();
-            txtTenPhong.Text = dGVDSOTro.Rows[i].Cells[7].Value.ToString();
-            txtLoaiPhong.Text = dGVDSOTro.Rows[i].Cells[8].Value.ToString();
-            txtGia.Text = dGVDSOTro.Rows[i].Cells[9].Value.ToString();
-            txtHoTen.Text = dGVDSOTro.Rows[i].Cells[12].Value.ToString();
-            txtCMND.Text = dGVDSOTro.Rows[i].Cells[13].Value.ToString();
-            txtSoDienThoai.Text = dGVDSOTro.Rows[i].Cells[14].Value.ToString();
-            dateSinh.Text = dGVDSOTro.Rows[i].Cells[15].Value.ToString();
-            txtDiaChi.Text = dGVDSOTro.Rows[i].Cells[16].Value.ToString();
-            cbBoxGioiTinh.Text = dGVDSOTro.Rows[i].Cells[17].Value.ToString();
+            anhxa();
+            if (dGVDSOTro.RowCount > 1)
+            {
+                int i = dGVDSOTro.CurrentRow.Index;
+                txtTienDatCoc.Text = dGVDSOTro.Rows[i].Cells[1].Value.ToString();
+                dTPNgayThue.Text = dGVDSOTro.Rows[i].Cells[2].Value.ToString();
+                dTPNgayTra.Text = dGVDSOTro.Rows[i].Cells[3].Value.ToString();
+                txtTenPhong.Text = dGVDSOTro.Rows[i].Cells[7].Value.ToString();
+                txtLoaiPhong.Text = dGVDSOTro.Rows[i].Cells[8].Value.ToString();
+                txtGia.Text = dGVDSOTro.Rows[i].Cells[9].Value.ToString();
+                txtHoTen.Text = dGVDSOTro.Rows[i].Cells[12].Value.ToString();
+                txtCMND.Text = dGVDSOTro.Rows[i].Cells[13].Value.ToString();
+                txtSoDienThoai.Text = dGVDSOTro.Rows[i].Cells[14].Value.ToString();
+                dateSinh.Text = dGVDSOTro.Rows[i].Cells[15].Value.ToString();
+                txtDiaChi.Text = dGVDSOTro.Rows[i].Cells[16].Value.ToString();
+                cbBoxGioiTinh.Text = dGVDSOTro.Rows[i].Cells[17].Value.ToString();
+            }
+            else
+            {
+                MessageBox.Show("Không có dữ liệu để chọn");
+            }
         }
         BusinessLogicLayer.BLLQLDSTro BLLQLDSTro = new BusinessLogicLayer.BLLQLDSTro();
         private void btnSua_Click(object sender, EventArgs e)
         {
-            int i = dGVDSOTro.CurrentRow.Index;
-            BLLQLDSTro.KTtext(txtTienDatCoc.Text, dTPNgayThue.Text, dTPNgayTra.Text, txtHoTen.Text, txtCMND.Text, txtSoDienThoai.Text, dateSinh.Text, txtDiaChi.Text, cbBoxGioiTinh.Text,txtTenPhong.Text, dGVDSOTro.Rows[i].Cells[12].Value.ToString());
+            if (dGVDSOTro.RowCount > 1) {
+                int i = dGVDSOTro.CurrentRow.Index;
+                BLLQLDSTro.KTtext(txtTienDatCoc.Text, dTPNgayThue.Text, dTPNgayTra.Text, txtHoTen.Text, txtCMND.Text, txtSoDienThoai.Text, dateSinh.Text, txtDiaChi.Text, cbBoxGioiTinh.Text,txtTenPhong.Text, dGVDSOTro.Rows[i].Cells[11].Value.ToString());
+            }
+            else
+            {
+                MessageBox.Show("Chưa có Khách Hàng");
+            }
             LoadDataGirView();
+            anhxa();
         }
 
         private void btnXoa_Click(object sender, EventArgs e)
         {
-            int i = dGVDSOTro.CurrentRow.Index;
-            //MessageBox.Show(""+dGVDSOTro.Rows[i].Cells[0].Value.ToString());
-            BLLQLDSTro.DeleteThuePhong(dGVDSOTro.Rows[i].Cells[0].Value.ToString(), dGVDSOTro.Rows[i].Cells[5].Value.ToString(), dGVDSOTro.Rows[i].Cells[11].Value.ToString());
+            if (dGVDSOTro.RowCount > 1) {
+                int i = dGVDSOTro.CurrentRow.Index;
+                BLLQLDSTro.DeleteThuePhong(dGVDSOTro.Rows[i].Cells[0].Value.ToString(), dGVDSOTro.Rows[i].Cells[5].Value.ToString(), dGVDSOTro.Rows[i].Cells[11].Value.ToString());
+            }
+            else
+            {
+               
+            }
+            anhxa();
             LoadDataGirView();
         }
